+++ src/cucu/cli/core.py
@@ -33,6 +33,7 @@ from cucu.cli import thread_dumper
 from cucu.cli.run import behave, behave_init, create_run
 from cucu.cli.steps import print_human_readable_steps, print_json_steps
 from cucu.config import CONFIG
+from cucu.db import consolidate_database_files, finish_worker_record
 from cucu.lint import linter
 from cucu.utils import generate_short_id
 
@@ -492,6 +493,10 @@ def run(
         if dumper is not None:
             dumper.stop()
 
+        if not dry_run and os.path.exists(results):
+            finish_worker_record()
+            consolidate_database_files(results)
+
         if generate_report:
             _generate_report(
                 results,
@@ -523,6 +528,9 @@ def _generate_report(
         shutil.rmtree(output)
 
     os.makedirs(output)
+
+    if os.path.exists(results_dir):
+        consolidate_database_files(results_dir)
 
     report_location = reporter.generate(
         results_dir, output, only_failures=only_failures